+++ indy_hub/__init__.py
@@ -2,4 +2,4 @@
 Indy Hub - An industrial management application for Alliance Auth
 """
 
-__version__ = "1.13.5"
+__version__ = "1.13.6"